Feature/create node

+++ src/modules/controllers/purchase.controllers.js
@@ -8,22 +8,43 @@ module.exports.getAllPurchase = (req, res) => {
 
 module.exports.addNewPurchase = (req, res) => {
   const purchase = new Purchase(req.body);
-  
-  purchase.save(req.body).then(result => {
-    res.send(result);
-  });
+  const body = req.body;
+
+  if (body.hasOwnProperty('text') 
+  && body.hasOwnProperty('date') 
+  && body.hasOwnProperty('cost')) {
+    purchase.save(req.body).then(() => {
+      Purchase.find().then(result => {
+        res.send({data: result});
+      });
+    });
+  } else {
+    res.status(422).send('Error!');
+  }
 };
 
 module.exports.deletePurchase = (req, res) => {
-  Purchase.deleteOne({ _id: req.query._id }).then(() => {
-    Purchase.find().then(result => {
-      res.send({data: result});
+  if (req.query._id) {
+    Purchase.deleteOne({ _id: req.query._id }).then(() => {
+      Purchase.find().then(result => {
+        res.send({data: result});
+      });
     });
-  });
+  } else {
+    res.status(422).send('Error!');
+  }
 };
 
 module.exports.editPurchase = (req, res) => {
-  Purchase.updateOne({ _id: req.body._id }, req.body).then(result => {
-    res.send({data: result});
-  });
+  const body = req.body;
+
+  if (body.hasOwnProperty('_id') && body._id) {
+    Purchase.updateOne({ _id: req.body._id }, req.body).then(() => {
+      Purchase.find().then(result => {
+        res.send({data: result});
+      });
+    });
+  } else {
+    res.status(422).send('Error!');
+  }
 };
